Fixes of formatted output.
git-svn-id: https://svn.r-project.org/R/trunk@85614 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/parallel/src/ncpus.c
@@ -87,8 +87,8 @@ static Rboolean ncpus_ex(int *ians)
                 if (!buffer) error(_("memory allocation error"));
 
             } else
-		error("in reading processor information, probable cause: %d",
-		      GetLastError());
+		error("in reading processor information, probable cause: %lu",
+		      (unsigned long)GetLastError());
         } else
 	    done = TRUE;
     }
@@ -122,8 +122,8 @@ static Rboolean ncpus_ex(int *ians)
 		         malloc(returnLength);
                 if (!buffer) error(_("memory allocation error"));
             } else
-		error("in reading processor information, probable cause: %d",
-		      GetLastError());
+		error("in reading processor information, probable cause: %lu",
+		      (unsigned long)GetLastError());
         } else done = TRUE;
     }
 
@@ -176,7 +176,9 @@ SEXP ncpus(SEXP virtual)
             if (GetLastError() == ERROR_INSUFFICIENT_BUFFER) {
                 buffer = (PSYSTEM_LOGICAL_PROCESSOR_INFORMATION) malloc(returnLength);
                 if (!buffer) error(_("memory allocation error"));
-            } else error("in reading processor information, probable cause: %d", GetLastError());
+            } else
+		error("in reading processor information, probable cause: %lu",
+		      (unsigned long)GetLastError());
         } else done = TRUE;
     }
